contract cleanup
